# entitlement-service/entitlement-service-openapi.yml
openapi: 3.0.3
info:
  title: Entitlement Service API
  description: |
    NextGen Workflow Entitlement Service - Central authority for authentication, authorization, and user management.
    
    This service provides:
    - Session-based authentication
    - Cerbos policy-based authorization
    - User and role management
    - Business application management
    
    **Authentication**: All endpoints except `/api/auth/login` require session-based authentication via `X-Session-Id` header.
    
    **Authorization**: Fine-grained access control is provided through Cerbos policy engine integration.
  version: 1.0.0
  contact:
    name: NextGen Workflow Team
servers:
  - url: http://localhost:8081
    description: Development server
...
paths:
...
  /api/entitlements/business-applications:
    get:
      tags:
        - Business Applications
      summary: Get all business applications
      description: Retrieve all business applications
      responses:
        '200':
          description: Business applications retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessApplication'

    post:
      tags:
        - Business Applications
      summary: Create business application
      description: Create a new business application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessApplication'
      responses:
        '201':
          description: Business application created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessApplication'
        '400':
          description: Invalid business application data
  /api/entitlements/business-applications/{id}:
    get:
      tags:
        - Business Applications
      summary: Get business application by ID
      description: Retrieve a specific business application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business application ID
      responses:
        '200':
          description: Business application found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessApplication'
        '404':
          description: Business application not found

    put:
      tags:
        - Business Applications
      summary: Update business application
      description: Update an existing business application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business application ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessApplication'
      responses:
        '200':
          description: Business application updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessApplication'
        '404':
          description: Business application not found

    delete:
      tags:
        - Business Applications
      summary: Delete business application
      description: Delete a business application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business application ID
      responses:
        '204':
          description: Business application deleted successfully
        '404':
          description: Business application not found
  /api/entitlements/business-applications/name/{businessAppName}:
    get:
      tags:
        - Business Applications
      summary: Get business application by name
      description: Retrieve a business application by name
      parameters:
        - name: businessAppName
          in: path
          required: true
          schema:
            type: string
          description: Business application name
      responses:
        '200':
          description: Business application found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessApplication'
        '404':
          description: Business application not found
...
  /api/entitlements/business-applications/{id}/activate:
    patch:
      tags:
        - Business Applications
      summary: Activate business application
      description: Activate a business application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business application ID
      responses:
        '200':
          description: Business application activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessApplication'
        '404':
          description: Business application not found
  /api/entitlements/business-applications/{id}/deactivate:
    patch:
      tags:
        - Business Applications
      summary: Deactivate business application
      description: Deactivate a business application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business application ID
      responses:
        '200':
          description: Business application deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessApplication'
        '404':
          description: Business application not found

  # Business App Roles endpoints
...
  /api/entitlements/business-app-roles/business-application/{businessAppId}:
    get:
      tags:
        - Business App Roles
      summary: Get roles by business application
      description: Retrieve all roles for a specific business application
      parameters:
        - name: businessAppId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business application ID
      responses:
        '200':
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessAppRole'
  /api/entitlements/business-app-roles/business-application/{businessAppId}/active:
    get:
      tags:
        - Business App Roles
      summary: Get active roles by business application
      description: Retrieve all active roles for a specific business application
      parameters:
        - name: businessAppId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business application ID
      responses:
        '200':
          description: Active roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessAppRole'
...
components:
...
  schemas:
...
    BusinessApplication:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        businessAppName:
          type: string
          example: "onecms"
        description:
          type: string
          example: "OneCMS Case Management System"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        metadata:
          type: object
          additionalProperties: true
          example:
            version: "1.0.0"
            owner: "Case Management Team"
    BusinessAppRole:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        businessApplication:
          $ref: '#/components/schemas/BusinessApplication'
        roleName:
          type: string
          example: "INTAKE_ANALYST"
        roleDisplayName:
          type: string
          example: "Intake Analyst"
        description:
          type: string
          example: "Analyst responsible for case intake and initial review"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        metadata:
          type: object
          additionalProperties: true
          example:
            permissions: ["READ", "CREATE"]
            queue: "intake-analyst-queue"
...